Rename MetarParser to metarParser
