+++ src/instruments/src/EFB/Performance/Widgets/LandingWidget.tsx
@@ -22,7 +22,7 @@ import { toast } from 'react-toastify';
 import { useSimVar } from '@instruments/common/simVars';
 import { MetarParserType } from '@instruments/common/metarTypes';
 import { Units } from '@shared/units';
-import { MetarParser } from '../../Utils/MetarParser';
+import { parseMetar } from '../../Utils/parseMetar';
 import { LandingCalculator, LandingFlapsConfig, LandingRunwayConditions } from '../Calculators/LandingCalculator';
 import RunwayVisualizationWidget, { LabelType } from './RunwayVisualizationWidget';
 import { SimpleInput } from '../../UtilComponents/Form/SimpleInput/SimpleInput';
@@ -78,7 +78,6 @@ export const LandingWidget = () => {
         flaps,
         runwayCondition,
         reverseThrust,
-        autoland,
         altitude,
         slope,
         temperature,
@@ -111,7 +110,6 @@ export const LandingWidget = () => {
             slope ?? 0,
             overweightProcedure,
             pressure ?? 0,
-            autoland,
         );
 
         dispatch(setLandingValues({
@@ -272,12 +270,6 @@ export const LandingWidget = () => {
         const reverseThrust: boolean = newValue;
 
         dispatch(setLandingValues({ reverseThrust }));
-    };
-
-    const handleAutolandChange = (newValue: boolean): void => {
-        const autoland: boolean = newValue;
-
-        dispatch(setLandingValues({ autoland }));
     };
 
     const handleRunwaySlopeChange = (value: string): void => {
@@ -623,18 +615,6 @@ export const LandingWidget = () => {
                                         defaultValue={initialState.landing.reverseThrust}
                                         value={reverseThrust}
                                         onChange={handleReverseThrustChange}
-                                        options={[
-                                            { value: false, displayValue: 'No' },
-                                            { value: true, displayValue: 'Yes' },
-                                        ]}
-                                    />
-                                </Label>
-                                <Label text="Autoland">
-                                    <SelectInput
-                                        className="w-64"
-                                        defaultValue={initialState.landing.autoland}
-                                        value={reverseThrust}
-                                        onChange={handleAutolandChange}
                                         options={[
                                             { value: false, displayValue: 'No' },
                                             { value: true, displayValue: 'Yes' },
